skipping save to HeadId collection with RANDOM commitId generator
