+++ javers-persistence-mongo/src/test/groovy/org/javers/repository/mongo/JaversMongoRepositoryE2ETest.groovy
@@ -46,33 +46,6 @@ abstract class JaversMongoRepositoryE2ETest extends JaversRepositoryShadowE2ETes
         snapshots[0].getPropertyValue('mapOfPrimitives') == ['primitive.value':1]
     }
 
-    def "should persist head id"() {
-        given:
-        MongoRepository mongoRepository = (MongoRepository)repository
-
-        def commitFactory = javersTestBuilder.commitFactory
-
-        def kazikV1 = dummyUser("Kazik").withAge(1)
-        def kazikV2 = dummyUser("Kazik").withAge(2)
-
-        def commit1 = commitFactory.create("author", [:], kazikV1)
-        def commit2 = commitFactory.create("author", [:], kazikV2)
-
-        when:
-        mongoRepository.persist(commit1)
-
-        then:
-        mongoRepository.getHeadId().getMajorId() == 1
-        mongoRepository.getHeadId().getMinorId() == 0
-
-        when:
-        mongoRepository.persist(commit2)
-
-        then:
-        mongoRepository.getHeadId().getMajorId() == 1
-        mongoRepository.getHeadId().getMinorId() == 1
-    }
-
     def "should persist commit and get latest snapshot"() {
         given:
         MongoRepository mongoRepository = (MongoRepository)repository
