+++ javers-spring-boot-starter-mongo/src/test/groovy/org/javers/spring/boot/mongo/JaversConcurrentTransactionsMongoTest.groovy
@@ -2,19 +2,21 @@ package org.javers.spring.boot.mongo
 
 import com.github.silaev.mongodb.replicaset.MongoDbReplicaSet
 import org.javers.core.Javers
-import org.javers.spring.transactions.JaversTransactionalTest
 import org.javers.spring.transactions.UberService
 import org.javers.spring.transactions.User
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
+import spock.lang.Specification
 
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 
-@SpringBootTest(classes = [TestApplication])
+@SpringBootTest(
+        classes = [TestApplication],
+        properties = ["javers.commitIdGenerator=random"])
 @ActiveProfiles("TransactionalMongo")
-class JaversMultipleTransactionsMongoTest extends JaversTransactionalTest {
+class JaversConcurrentTransactionsMongoTest extends Specification {
 
     @Autowired
     MongoDbReplicaSet replicaSet
@@ -28,7 +30,7 @@ class JaversMultipleTransactionsMongoTest extends JaversTransactionalTest {
     @Autowired
     private UberService uberService
 
-    def "should not produce write conflict on concurrent commits"() {
+    def "should not produce write conflict on concurrent commits with RANDOM CommitId generator"() {
         given:
         def tasks = [
                 { -> javers.commit("author_1", new User()) },
